Fix line endings for full-width boxes
When running boxen in an environment in which the console columns cannot be detected, the column width defaults to 80 and the box width is capped at this same amount. The actual display width however, can still be greater (or even unlimited when piping to a file). Therefore we still need to use newlines when the box width equals the columns width.